+++ CHANGELOG.md
@@ -34,6 +34,8 @@ Versioning].
   local offset at the provided moment in time.
   - `OffsetDateTime::now()` and `impl From<SystemTime> for OffsetDateTime` both
     return a value with the local offset.
+  - `OffsetDateTime::now_utc()` is equivalent to calling
+    `OffsetDateTime::now().to_offset(offset!(UTC))`.
   - `UtcOffset::current_local_offset()` will return the equivalent of
     `OffsetDateTime::now().offset()`.